requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pydantic
sqlalchemy
passlib[bcrypt]
python-jose
scikit-learn
pandas
numpy
requests
email-validator
